requirements.txt
# note this is required
#   pip install --upgrade "pip>=19.3"
# then
#   pip install -r requirements.txt

absl-py
genanki
